requirements.txt
numpy==1.24.1
opencv_contrib_python==4.7.0.72
opencv_python==4.6.0.66
Pillow==9.3.0
Pillow==9.5.0
pytesseract==0.3.10
tesseract.exe if pytessaract gives an error
